+++ .depreciated-iter2/postman_examples/AWS IOT Testing Environment.postman_environment.json
