+++ .depreciated-iter2/scripts/authprocess.js
