+++ .depreciated-iter2/scripts/authworker.js
