+++ .depreciated-iter2/scripts/logout.js
