+++ .depreciated-iter2/scripts/services/auth-service.js
