+++ .depreciated-iter2/views/pages/sync.html
