+++ .depreciated-iter2/views/styles/settings.css
